Desafio 14 e 15 adicionados e demais exercícios corrigidos

+++ AulasCursoEmVideo/aulas/aula07/aula07Desafio07.py
@@ -1,5 +1,5 @@
-nota1 = int(input('Digite a nota 1: '))
-nota2 = int(input('Digite a nota 2: '))
+nota1 = float(input('Digite a nota 1: '))
+nota2 = float(input('Digite a nota 2: '))
 
 media = (nota1 + nota2)/2
